fix: 빌드 오류 해결을 위한 authOption 폴더 구조 리팩토링
