+++ app/lib/index.tsx
@@ -1,1 +1,2 @@
 export * from './Provider';
+export * from './authOptions';
